Move markers to runtime
