+++ circuit-runtime/src/commonMain/kotlin/com/slack/circuit/runtime/markers.kt
@@ -1,6 +1,6 @@
 // Copyright (C) 2022 Slack Technologies, LLC
 // SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit
+package com.slack.circuit.runtime
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
