+++ circuit-test/src/commonMain/kotlin/com/slack/circuit/test/PresenterTestExtensions.kt
@@ -7,7 +7,7 @@ import app.cash.molecule.RecompositionClock
 import app.cash.molecule.moleculeFlow
 import app.cash.turbine.ReceiveTurbine
 import app.cash.turbine.test
-import com.slack.circuit.CircuitUiState
+import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.Presenter
 import kotlin.time.Duration
