+++ circuit/src/commonMain/kotlin/com/slack/circuit/CircuitContent.kt
@@ -8,6 +8,7 @@ import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import com.slack.circuit.runtime.CircuitUiState
 
 @Composable
 public fun CircuitContent(
