+++ circuit/src/commonMain/kotlin/com/slack/circuit/Ui.kt
@@ -4,6 +4,7 @@ package com.slack.circuit
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.slack.circuit.runtime.CircuitUiState
 
 /**
  * Represents a composable UI for the given [UiState]. Conventionally, this should just be the
